Refactoring Namespaces; Rename from ByJG\Swagger to ByJG\ApiTools;
