+++ src/OpenApi/OpenApiResponseBody.php
@@ -1,10 +1,11 @@
 <?php
 
-namespace ByJG\Swagger;
+namespace ByJG\ApiTools\OpenApi;
 
-use ByJG\Swagger\Exception\NotMatchedException;
+use ByJG\ApiTools\Exception\NotMatchedException;
+use ByJG\ApiTools\SwaggerBody;
 
-class SwaggerResponseBody extends SwaggerBody
+class OpenApiResponseBody extends SwaggerBody
 {
     /**
      * @param $body
@@ -17,12 +18,12 @@ class SwaggerResponseBody extends SwaggerBody
      */
     public function match($body)
     {
-        if (!isset($this->structure['schema'])) {
+        if (!isset($this->structure['content'])) {
             if (!empty($body)) {
                 throw new NotMatchedException("Expected empty body for " . $this->name);
             }
             return true;
         }
-        return $this->matchSchema($this->name, $this->structure['schema'], $body);
+        return $this->matchSchema($this->name, $this->structure['content'][key($this->structure['content'])]['schema'], $body);
     }
 }
