+++ tests/SwaggerBodyTestCase.php
@@ -2,7 +2,7 @@
 
 namespace Test;
 
-use ByJG\Swagger\SwaggerSchema;
+use ByJG\ApiTools\Swagger\SwaggerSchema;
 use PHPUnit\Framework\TestCase;
 
 class SwaggerBodyTestCase extends TestCase
@@ -14,7 +14,7 @@ class SwaggerBodyTestCase extends TestCase
      */
     protected static function swaggerSchema($allowNullValues = false)
     {
-        return \ByJG\Swagger\Base\Schema::getInstance(
+        return \ByJG\ApiTools\Base\Schema::getInstance(
             self::getSwaggerJsonContent(),
             $allowNullValues
         );
@@ -26,7 +26,7 @@ class SwaggerBodyTestCase extends TestCase
      */
     protected static function swaggerSchema2($allowNullValues = false)
     {
-        return \ByJG\Swagger\Base\Schema::getInstance(
+        return \ByJG\ApiTools\Base\Schema::getInstance(
             self::getSwaggerJsonContent_No2(),
             $allowNullValues
         );
